backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-community
langchain-text-splitters
langchain-groq
langchain
pinecone-client
PyMuPDF
sentence-transformers
requests
groq
